+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enabled automatic data gathering for custom currencies added via the currency management in the admin control panel
 
+### Changed
+
+- Refactored the create or update access dialog component to standalone
+
 ### Fixed
 
 - Fixed an issue related to the error handling in the data provider status component